backend/pyproject.toml
[project]
name = "lcms-backend"
version = "1.0.0"
description = "Learning Content Management System - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "sqlalchemy[asyncio]==2.0.30",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "python-jose[cryptography]==3.3.0",
    "bcrypt==4.1.3",
    "python-multipart==0.0.9",
    "boto3==1.34.84",
    "httpx==0.27.0",
    "email-validator==2.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "pytest-cov==5.0.0",
    "httpx==0.27.0",
    "faker==24.11.0",
    "ruff==0.4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]
